[lab3] add code for the rectangle drawing

+++ lab3/Landscape/Common/Types/DimensionTypes.h
@@ -2,27 +2,27 @@
 
 struct Size
 {
-	double width;
-	double height;
+	float width;
+	float height;
 };
 
 struct Vector2d
 {
-	double x;
-	double y;
+	float x;
+	float y;
 };
 
 struct Vector3d
 {
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 };
 
 struct Vector4d
 {
-	double x;
-	double y;
-	double z;
-	double w;
+	float x;
+	float y;
+	float z;
+	float w;
 };
